backends: fix double-promotion warnings

+++ backends/gf-monitor-config-store.c
@@ -1000,11 +1000,11 @@ handle_text (GMarkupParseContext  *context,
                            &parser->current_logical_monitor_config->scale, error))
             return;
 
-          if (parser->current_logical_monitor_config->scale <= 0.0)
+          if (parser->current_logical_monitor_config->scale <= 0.0f)
             {
               g_set_error (error, G_IO_ERROR, G_IO_ERROR_FAILED,
                            "Logical monitor scale '%g' invalid",
-                           parser->current_logical_monitor_config->scale);
+                           (gdouble) parser->current_logical_monitor_config->scale);
               return;
             }
